+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 * Allow category map to be passed as YAML in cli modify-categories
+* Better support for confusion kwcoco and visualize script in evaluate detections
 
 
 ## Version 0.8.8 - Released 2025-07-15